Rework the API to be idempotent
login/logout happens automatically upon any operation. This should eliminate a full class of possible bugs, and make things generally more stable and understandable.
This was done in part to fix the toggle_sticky api. It wasn't working due to the 2nd login being ignored by the server, but the session id of the new login (which is worthless) was saved and used instead.

This does require wainting 3 seconds between each API call, due to the forum implementing a lockout in between logins.
However, since the alternative is managing and saving a single session, tracking if it's still alive, and having error handling in case it dies at an unexpected time, this is much simpler and more reliable.

This also added error checking to verify we login/logout.